requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
sounddevice>=0.4.7
numpy>=1.26.0
